add graphical data format endpoint and refactor
